+++ cpp/src/arrow/builder_benchmark.cc
@@ -337,9 +337,9 @@ static void BuildStringDictionaryArray(
     benchmark::State& state) {  // NOLINT non-const reference
   const auto fodder = MakeStringDictFodder();
   auto fodder_nbytes =
-      std::accumulate(fodder.begin(), fodder.end(), 0ULL,
+      std::accumulate(fodder.begin(), fodder.end(), static_cast<size_t>(0ULL),
                       [&](size_t acc, const std::string& s) { return acc + s.size(); });
-  BenchmarkDictionaryArray<BinaryDictionaryBuilder>(state, fodder, fodder_nbytes);
+  BenchmarkDictionaryArray<BinaryDictionaryBuilder>(state, fodder, static_cast<size_t>(fodder_nbytes));
 }
 
 static void ArrayDataConstructDestruct(
